+++ benchflow-test-manager/application/src/test/java/cloud/benchflow/testmanager/services/internal/dao/BenchFlowExperimentModelDAOIT.java
@@ -1,13 +1,13 @@
 package cloud.benchflow.testmanager.services.internal.dao;
 
 import static cloud.benchflow.testmanager.constants.BenchFlowConstants.MODEL_ID_DELIMITER;
-import static cloud.benchflow.testmanager.helpers.TestConstants.LOAD_TEST_NAME;
+import static cloud.benchflow.testmanager.helpers.constants.TestConstants.LOAD_TEST_NAME;
 import static org.junit.Assert.assertEquals;
 
 import cloud.benchflow.faban.client.responses.RunStatus;
 import cloud.benchflow.testmanager.DockerComposeIT;
 import cloud.benchflow.testmanager.exceptions.BenchFlowExperimentIDDoesNotExistException;
-import cloud.benchflow.testmanager.helpers.TestConstants;
+import cloud.benchflow.testmanager.helpers.constants.TestConstants;
 import cloud.benchflow.testmanager.models.BenchFlowExperimentModel;
 import cloud.benchflow.testmanager.models.BenchFlowTestModel;
 import cloud.benchflow.testmanager.models.User;
