+++ benchflow-test-manager/application/src/test/java/cloud/benchflow/testmanager/strategy/selection/OneAtATimeSelectionStrategyTest.java
@@ -1,7 +1,7 @@
 package cloud.benchflow.testmanager.strategy.selection;
 
-import cloud.benchflow.testmanager.helpers.TestConstants;
-import cloud.benchflow.testmanager.helpers.TestFiles;
+import cloud.benchflow.testmanager.helpers.constants.TestConstants;
+import cloud.benchflow.testmanager.helpers.constants.TestFiles;
 import cloud.benchflow.testmanager.services.external.MinioService;
 import cloud.benchflow.testmanager.services.internal.dao.ExplorationModelDAO;
 import cloud.benchflow.testmanager.strategy.selection.SelectionStrategy.SelectedExperimentBundle;
